refactor: create encryption feature
