+++ src/core/features/encryption/workers/Cryptography.worker.ts
@@ -1,14 +1,15 @@
 import { WorkerMessenger } from '@utils/workers/WorkerMessenger';
 import { WorkerRPC } from '@utils/workers/WorkerRPC';
 
-import { AESGCMCipher } from '../encryption/ciphers/AES';
-import { TwofishCTRCipher } from '../encryption/ciphers/Twofish';
-import { EncryptionController } from '../encryption/EncryptionController';
-import { BufferIntegrityProcessor } from '../encryption/processors/BufferIntegrityProcessor';
-import { BufferSizeObfuscationProcessor } from '../encryption/processors/BufferSizeObfuscationProcessor';
-import { PipelineProcessor } from '../encryption/processors/PipelineProcessor';
-import { getDerivedKeysManager, getMasterKey } from '../encryption/utils/keys';
-import { getRandomBytes } from '../encryption/utils/random';
+import { AESGCMCipher } from '../../../encryption/ciphers/AES';
+import { TwofishCTRCipher } from '../../../encryption/ciphers/Twofish';
+import { EncryptionController } from '../../../encryption/EncryptionController';
+import { BufferIntegrityProcessor } from '../../../encryption/processors/BufferIntegrityProcessor';
+import { BufferSizeObfuscationProcessor } from '../../../encryption/processors/BufferSizeObfuscationProcessor';
+import { PipelineProcessor } from '../../../encryption/processors/PipelineProcessor';
+import { getDerivedKeysManager, getMasterKey } from '../../../encryption/utils/keys';
+import { getRandomBytes } from '../../../encryption/utils/random';
+
 import { FakeWorkerObject } from '.';
 
 export default FakeWorkerObject;
