+++ src/core/features/encryption/workers/index.ts
