+++ src/state/profiles.ts
@@ -4,9 +4,9 @@ import { useUnit } from 'effector-react';
 import path from 'path';
 import { EncryptionController } from '@core/encryption/EncryptionController';
 import { PlaceholderEncryptionController } from '@core/encryption/PlaceholderEncryptionController';
-import { createEncryption } from '@core/encryption/utils/createEncryption';
 import { base64ToBytes } from '@core/encryption/utils/encoding';
 import { AttachmentsController } from '@core/features/attachments/AttachmentsController';
+import { createEncryption } from '@core/features/encryption/createEncryption';
 import { FilesController } from '@core/features/files/FilesController';
 import { NotesController } from '@core/features/notes/controller/NotesController';
 import { TagsController } from '@core/features/tags/controller/TagsController';
