+++ src/state/profilesManager.ts
@@ -2,9 +2,9 @@ import { useCallback, useEffect, useState } from 'react';
 import { createEvent, createStore } from 'effector';
 import { useUnit } from 'effector-react';
 import { EncryptionController } from '@core/encryption/EncryptionController';
-import { WorkerEncryptionProxyProcessor } from '@core/encryption/processors/WorkerEncryptionProxyProcessor';
 import { bytesToBase64 } from '@core/encryption/utils/encoding';
 import { getRandomBytes } from '@core/encryption/utils/random';
+import { WorkerEncryptionProxyProcessor } from '@core/features/encryption/workers/WorkerEncryptionProxyProcessor';
 import { ProfileObject, ProfilesManager } from '@core/storage/ProfilesManager';
 import { ElectronFilesController } from '@electron/requests/storage/renderer';
 import { NewProfile } from '@features/WorkspaceManager/ProfileCreator';
@@ -64,6 +64,7 @@ export const useProfilesManager = () => {
 			// Create encrypted profile
 			const salt = getRandomBytes(96);
 
+			// TODO: replace to `createEncryption`
 			const workerEncryption = new WorkerEncryptionProxyProcessor(
 				profile.password,
 				new Uint8Array(salt),
